requirements.txt
PyPDF2
python-docx
pycryptodome
rapidfuzz>=2.1.3 